requirements.txt
langchain-google-genai
langgraph
crewai
langchain
pydantic
python-dotenv
streamlit
google-generativeai
pypdf
gTTS
fpdf